Add edge case for when airport storage is empty

+++ lib/airport.rb
@@ -5,7 +5,8 @@ class Airport
 
   DEFAULT_CAPACITY = 20
 
-  attr_accessor :storage, :weather, :capacity
+  attr_accessor :capacity
+  attr_reader :storage, :weather
 
   def initialize(weather = Weather.new, capacity = DEFAULT_CAPACITY)
     @storage = []
@@ -15,16 +16,27 @@ class Airport
 
   def land(plane)
     fail "Storm's a-brewing - Better divert course!" if weather.stormy?
-    fail "No room at the Inn - Try another airport!" if storage.length >= capacity
-    storage << plane
+    fail "No room at the Inn - Try another airport!" if full?
     plane.status = "Grounded"
-    "#{plane} has now landed."
+    puts "#{plane} has now landed."
+    storage << plane
   end
 
   def takeoff(plane)
     fail "Storm's a-brewing - Better hold tight!" if weather.stormy?
-    storage.delete(plane)
+    fail "No planes available" if empty?
     plane.status = "Airborne"
-    "#{plane} has now taken off"
+    puts "#{plane} has now taken off"
+    storage.delete(plane)
+  end
+
+  private
+
+  def full?
+    storage.length >= capacity
+  end
+
+  def empty?
+    storage.empty?
   end
 end
